+++ packages/workspace-tools/CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.162.1 (2024-08-27)
 
-
 ### Bug Fixes
 
-- **k8s-tools:** Resolved issue with invalid import ([82a782d4](https://github.com/storm-software/storm-ops/commit/82a782d4))
+- **k8s-tools:** Resolved issue with invalid import
+  ([82a782d4](https://github.com/storm-software/storm-ops/commit/82a782d4))
 
 ## 1.162.0 (2024-08-27)